Conflicts:
	Proftaak/Businesslayer/DAL/DbAccess.cs

+++ Proftaak/Businesslayer/Business/ReservationMaterial.cs
@@ -8,14 +8,14 @@ namespace Businesslayer.Business
     {
         DbMaterial DM = new DbMaterial();
         public int Amount { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
-       public ReservationMaterial(DateTime beginTime, DateTime endtime, User employee, int amount, decimal price, Item item) : base(beginTime, endtime, employee)          
+       public ReservationMaterial(DateTime beginTime, DateTime endtime, User employee, int amount, int price, Item item) : base(beginTime, endtime, employee)          
        {
            Amount = amount;
            Price = price;
        }
-        public void ChangePrice(string name, decimal price)
+        public void ChangePrice(string name, int price)
         {
           // DM.ChangePrice(name, price);        
         }
